requirements.txt
# requirements for basic usage
numpy
Pillow
PyYAML
torch
torchvision

# requirements for training
wandb

# requirements for video inference
opencv-python

# requirements for TensorRT usage
onnx
onnxruntime
pycuda
tensorrt